requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
psycopg2-binary
psycopg_binary
psycopg
django-environ
